Rename ParseOrderData.kt to FormatOrderData
